Rename the brittle test - prepare to remove it
