+++ src/test/java/hudson/plugins/git/util/GitUtilsTest.java
@@ -23,276 +23,236 @@
  */
 package hudson.plugins.git.util;
 
-import edu.umd.cs.findbugs.annotations.NonNull;
 import hudson.EnvVars;
+import hudson.model.TaskListener;
 import hudson.plugins.git.Branch;
 import hudson.plugins.git.BranchSpec;
 import hudson.plugins.git.Revision;
-import hudson.util.LogTaskListener;
 import hudson.util.StreamTaskListener;
-
 import java.io.File;
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collection;
-import java.util.Collections;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-
+import java.util.Random;
+import jenkins.plugins.git.GitSampleRepoRule;
 import org.eclipse.jgit.lib.ObjectId;
-
+import static org.hamcrest.Matchers.is;
+import static org.hamcrest.Matchers.nullValue;
 import org.jenkinsci.plugins.gitclient.Git;
 import org.jenkinsci.plugins.gitclient.GitClient;
+import static org.junit.Assert.assertThat;
 import org.junit.Before;
+import org.junit.BeforeClass;
+import org.junit.ClassRule;
+import org.junit.Rule;
 import org.junit.Test;
-import static org.hamcrest.Matchers.*;
-import static org.junit.Assert.*;
+import org.junit.rules.TemporaryFolder;
 
-/**
- * This test was a useful first draft test because it used the commit volume and
- * tag volume of the git plugin repository to quickly show the problem.
- *
- * This test is NOT a reasonable test to include in the final pull request
- * because it is brittle. It assumes the HEAD commit exists on the remote
- * origin. That is a reasonable assumption in many continuous integration
- * environments. That is not a reasonable assumption in a development
- * environment. These tests MUST be replaced.
- *
- * @author Mark Waite
- */
 public class GitUtilsTest {
 
+    @ClassRule
+    public static GitSampleRepoRule sampleOriginRepo = new GitSampleRepoRule();
+
+    @Rule
+    public TemporaryFolder repoParentFolder = new TemporaryFolder();
+
+    @Rule
+    public GitSampleRepoRule sampleRepo = new GitSampleRepoRule();
+
+    private static final String[] BRANCH_NAMES = {
+        "master",
+        "sally-2",
+        "baker-1",
+        "able-4"
+    };
+    private static final String OLDER_BRANCH_NAME = "older-branch";
+
+    private static ObjectId headId = null;
+    private static ObjectId priorHeadId = null;
+
+    private static Revision headRevision = null;
+    private static Revision priorRevision = null;
+
+    private static final String PRIOR_TAG_NAME_1 = "prior-tag-1";
+    private static final String PRIOR_TAG_NAME_2 = "prior-tag-2-annotated";
+    private static final String HEAD_TAG_NAME_1 = "head-tag-1";
+    private static final String HEAD_TAG_NAME_2 = "head-tag-2-annotated";
+
+    private static List<BranchSpec> branchSpecList = null;
+    private static List<BranchSpec> priorBranchSpecList = null;
+    private static List<Branch> branchList = null;
+
     private final EnvVars env;
-    private final File gitDir;
-    private final ObjectId remoteHeadId;
-    private final ObjectId remotePriorId;
-    private final Set<Branch> branches;
-    private final Set<String> tagNames;
-    private final String branchName;
-
-    private final List<BranchSpec> currentBranchSpecList;
-    private final Revision remoteHeadRevision;
-    private final Revision remotePriorRevision;
-
-    private LogHandler handler;
-    private LogTaskListener listener;
-    private int logCount = 0;
+    private final TaskListener listener = StreamTaskListener.NULL;
 
     private GitUtils gitUtils;
     private GitClient gitClient;
 
-    private static final String LOGGING_STARTED = "*** Logging started ***";
-    private final List<String> expectedLogSubstrings;
-    private final List<String> unexpectedLogSubstrings;
+    private static final Random RANDOM = new Random();
 
-    public GitUtilsTest() throws IOException, InterruptedException {
+    public GitUtilsTest() {
         this.env = new EnvVars();
+    }
 
-        String gitDirName = ".";
-        this.gitDir = new File(gitDirName);
+    @BeforeClass
+    public static void createSampleOriginRepo() throws Exception {
+        String fileName = "README";
+        sampleOriginRepo.init();
+        sampleOriginRepo.git("config", "user.name", "Author User Name");
+        sampleOriginRepo.git("config", "user.email", "author.user.name@mail.example.com");
+        sampleOriginRepo.git("tag", PRIOR_TAG_NAME_1);
+        sampleOriginRepo.git("tag", "-a", PRIOR_TAG_NAME_2, "-m", "Annotated tag " + PRIOR_TAG_NAME_2);
+        priorHeadId = ObjectId.fromString(sampleOriginRepo.head());
 
-        /* Assemble information about the repository in this directory */
-        GitClient oneTimeGitClient = Git.with(StreamTaskListener.NULL, env).in(gitDir).using("git").getClient();
+        sampleOriginRepo.git("checkout", "-b", OLDER_BRANCH_NAME);
+        branchList = new ArrayList<>();
+        branchList.add(new Branch(OLDER_BRANCH_NAME, priorHeadId));
+        branchList.add(new Branch("refs/tags/" + PRIOR_TAG_NAME_1, priorHeadId));
+        branchList.add(new Branch("refs/tags/" + PRIOR_TAG_NAME_2, priorHeadId));
+        priorRevision = new Revision(priorHeadId, branchList);
+        priorBranchSpecList = new ArrayList<>();
+        priorBranchSpecList.add(new BranchSpec(OLDER_BRANCH_NAME));
 
-        /* Tag names and remote branch names */
-        this.tagNames = Collections.unmodifiableSet(oneTimeGitClient.getTagNames(null));
-        // assertFalse("No tags in this repository - reduces test strength", this.tagNames.isEmpty());
-        Set<Branch> remoteBranches = oneTimeGitClient.getRemoteBranches();
-        this.branches = Collections.unmodifiableSet(remoteBranches);
-        // assertFalse("No branches in this repository - reduces test strength", this.branches.isEmpty());
-        String guessedBranchName = "HEAD"; // In case local branch matches no remote branch
-        ObjectId headId = oneTimeGitClient.revParse("HEAD");
-        for (Branch branch : this.branches) {
-            // Choose one of the remote branch names that matches the current checked out version
-            if (headId.equals(branch.getSHA1()) && !branch.getName().endsWith("/HEAD")) {
-                guessedBranchName = branch.getName();
+        sampleOriginRepo.git("checkout", "master");
+        sampleOriginRepo.write(fileName, "This is the README file " + RANDOM.nextInt());
+        sampleOriginRepo.git("add", fileName);
+        sampleOriginRepo.git("commit", "-m", "Adding " + fileName, fileName);
+        sampleOriginRepo.git("tag", HEAD_TAG_NAME_1);
+        sampleOriginRepo.git("tag", "-a", HEAD_TAG_NAME_2, "-m", "Annotated tag " + HEAD_TAG_NAME_2);
+        headId = ObjectId.fromString(sampleOriginRepo.head());
+        branchSpecList = new ArrayList<>();
+        branchList = new ArrayList<>();
+        branchSpecList.add(new BranchSpec("master"));
+        branchSpecList.add(new BranchSpec("refs/tags/" + HEAD_TAG_NAME_1));
+        branchSpecList.add(new BranchSpec("refs/tags/" + HEAD_TAG_NAME_2));
+        branchList.add(new Branch("master", headId));
+        branchList.add(new Branch("refs/tags/" + HEAD_TAG_NAME_1, headId));
+        branchList.add(new Branch("refs/tags/" + HEAD_TAG_NAME_2, headId));
+        for (String branchName : BRANCH_NAMES) {
+            if (!branchName.equals("master")) {
+                sampleOriginRepo.git("checkout", "-b", branchName);
+                branchSpecList.add(new BranchSpec(branchName));
+                branchList.add(new Branch(branchName, headId));
             }
         }
-        assertThat(guessedBranchName, not(isEmptyString()));
-        this.branchName = guessedBranchName;
-
-        /* Assumes the HEAD commit exists on the remote */
-        this.remoteHeadId = oneTimeGitClient.revParse(this.branchName);
-        this.remoteHeadRevision = new Revision(remoteHeadId);
-        this.remotePriorId = oneTimeGitClient.revParse(this.branchName + "^");
-        this.remotePriorRevision = new Revision(remotePriorId);
-
-        this.currentBranchSpecList = new ArrayList<>();
-        // BranchSpec headSpec = new BranchSpec("HEAD");
-        // currentBranchSpecList.add(headSpec);
-        BranchSpec currentBranchSpec = new BranchSpec(guessedBranchName);
-        currentBranchSpecList.add(currentBranchSpec);
-
-        this.expectedLogSubstrings = new ArrayList<>();
-        this.unexpectedLogSubstrings = new ArrayList<>();
+        sampleOriginRepo.git("checkout", "master"); // Master branch as current branch in origin repo
+        headRevision = new Revision(headId, branchList);
     }
 
     @Before
-    public void setUp() throws IOException, InterruptedException {
-        Logger logger = Logger.getLogger(this.getClass().getPackage().getName() + "-" + logCount++);
-        handler = new LogHandler();
-        handler.setLevel(Level.ALL);
-        logger.setUseParentHandlers(false);
-        logger.addHandler(handler);
-        logger.setLevel(Level.ALL);
-        this.listener = new hudson.util.LogTaskListener(logger, Level.ALL);
-        listener.getLogger().println(LOGGING_STARTED);
-
+    public void cloneSampleRepo() throws Exception {
+        File gitDir = repoParentFolder.newFolder("test-repo");
         this.gitClient = Git.with(listener, env).in(gitDir).using("git").getClient();
+        this.gitClient.init();
+        this.gitClient.clone_().url(sampleOriginRepo.fileUrl()).repositoryName("origin").execute();
+        this.gitClient.checkout("origin/master", "master");
         this.gitUtils = new GitUtils(listener, gitClient);
-    }
-
-    private Set<String> getExpectedNames(@NonNull Set<Branch> branches, @NonNull Set<String> tagNames) {
-        Set<String> names = new HashSet<>(branches.size() + tagNames.size());
-        for (Branch branch : branches) {
-            names.add(branch.getName());
-        }
-        for (String tagName : tagNames) {
-            names.add("refs/tags/" + tagName);
-        }
-        return names;
-    }
-
-    private Set<String> getExpectedNames(@NonNull List<BranchSpec> branchSpecList) {
-        Set<String> names = new HashSet<>(branchSpecList.size());
-        for (BranchSpec branchSpec : branchSpecList) {
-            names.add(branchSpec.getName());
-        }
-        return names;
-    }
-
-    private Set<String> getActualNames(@NonNull Collection<Revision> revisions) {
-        Set<String> names = new HashSet<>(revisions.size());
-        for (Revision revision : revisions) {
-            for (Branch branch : revision.getBranches()) {
-                names.add(branch.getName());
-            }
-        }
-        return names;
-    }
-
-    @Test
-    public void testGetAllBranchRevisions() throws Exception {
-        Collection<Revision> allRevisions = gitUtils.getAllBranchRevisions();
-        assertThat(allRevisions, hasItem(remoteHeadRevision));
-        Set<String> expectedNames = getExpectedNames(branches, tagNames);
-        Set<String> actualNames = getActualNames(allRevisions);
-        assertThat(actualNames, is(expectedNames));
-    }
-
-    @Test
-    public void testGetMatchingRevisions() throws Exception {
-        Collection<Revision> allRevisions = gitUtils.getMatchingRevisions(currentBranchSpecList, env);
-        // Collection<Revision> allRevisions = gitUtils.getAllBranchRevisions();
-        assertThat(allRevisions, hasItem(remoteHeadRevision));
-        Set<String> actualNames = getActualNames(allRevisions);
-        Set<String> expectedNames = getExpectedNames(currentBranchSpecList);
-        assertThat(actualNames, is(expectedNames));
-    }
-
-    @Test
-    public void testGetMatchingRevisions_Duplicates() throws Exception {
-        List<BranchSpec> duplicateBranchSpecList = new ArrayList<>();
-        duplicateBranchSpecList.addAll(currentBranchSpecList);
-        duplicateBranchSpecList.addAll(currentBranchSpecList);
-        Collection<Revision> allRevisions = gitUtils.getMatchingRevisions(duplicateBranchSpecList, env);
-        // Collection<Revision> allRevisions = gitUtils.getAllBranchRevisions();
-        assertThat(allRevisions, hasItem(remoteHeadRevision));
-        Set<String> actualNames = getActualNames(allRevisions);
-        Set<String> expectedNames = getExpectedNames(currentBranchSpecList);
-        assertThat(actualNames, is(expectedNames));
-    }
-
-    @Test
-    public void testGetRevisionContainingBranch() throws Exception {
-        Revision revision = gitUtils.getRevisionContainingBranch(branchName);
-        assertEquals(remoteHeadRevision, revision);
-    }
-
-    @Test
-    public void testGetRevisionContainingBranchInvalidName() throws Exception {
-        Revision revision = gitUtils.getRevisionContainingBranch(branchName + "-invalid-name");
-        assertThat(revision, is(nullValue()));
-    }
-
-    @Test
-    public void testGetRevisionForSHA1() throws Exception {
-        Revision revision = gitUtils.getRevisionForSHA1(remoteHeadId);
-        assertEquals(remoteHeadRevision, revision);
-    }
-
-    @Test
-    public void testGetRevisionForSHA1PriorRevision() throws Exception {
-        Revision revision = gitUtils.getRevisionForSHA1(remotePriorId);
-        assertEquals(remotePriorRevision, revision);
     }
 
     @Test
     public void testSortBranchesForRevision_Revision_List() {
-        Revision result = gitUtils.sortBranchesForRevision(remoteHeadRevision, currentBranchSpecList);
-        assertEquals(remoteHeadRevision, result);
+        Revision result = gitUtils.sortBranchesForRevision(headRevision, branchSpecList);
+        assertThat(result, is(headRevision));
     }
 
     @Test
     public void testSortBranchesForRevision_Revision_List_Prior() {
-        Revision result = gitUtils.sortBranchesForRevision(remotePriorRevision, currentBranchSpecList);
-        assertEquals(remotePriorRevision, result);
+        Revision result = gitUtils.sortBranchesForRevision(priorRevision, priorBranchSpecList);
+        assertThat(result, is(priorRevision));
     }
 
     @Test
-    public void testSortBranchesForRevision_3args() {
-        Revision result = gitUtils.sortBranchesForRevision(remoteHeadRevision, currentBranchSpecList, env);
-        assertEquals(remoteHeadRevision, result);
+    public void testSortBranchesForRevision_Revision_List_Mix_1() {
+        Revision result = gitUtils.sortBranchesForRevision(headRevision, priorBranchSpecList);
+        assertThat(result, is(headRevision));
     }
 
     @Test
-    public void testSortBranchesForRevision_3args_Prior() {
-        Revision result = gitUtils.sortBranchesForRevision(remotePriorRevision, currentBranchSpecList, env);
-        assertEquals(remotePriorRevision, result);
+    public void testSortBranchesForRevision_Revision_List_Mix_2() {
+        Revision result = gitUtils.sortBranchesForRevision(priorRevision, branchSpecList);
+        assertThat(result, is(priorRevision));
+    }
+
+    @Test
+    public void testSortBranchesForRevision_Revision_List_Prior_3_args() {
+        Revision result = gitUtils.sortBranchesForRevision(headRevision, branchSpecList, env);
+        assertThat(result, is(headRevision));
+    }
+
+    @Test
+    public void testGetRevisionContainingBranch() throws Exception {
+        for (String branchName : BRANCH_NAMES) {
+            Revision revision = gitUtils.getRevisionContainingBranch("origin/" + branchName);
+            assertThat(revision, is(headRevision));
+        }
+    }
+
+    /* Tags are searched in getRevisionContainingBranch beginning with 3.2.0 */
+    @Test
+    public void testGetRevisionContainingBranch_UseTagNameHead1() throws Exception {
+        Revision revision = gitUtils.getRevisionContainingBranch("refs/tags/" + HEAD_TAG_NAME_1);
+        assertThat(revision, is(headRevision));
+    }
+
+    /* Tags are searched in getRevisionContainingBranch beginning with 3.2.0 */
+    @Test
+    public void testGetRevisionContainingBranch_UseTagNameHead2() throws Exception {
+        Revision revision = gitUtils.getRevisionContainingBranch("refs/tags/" + HEAD_TAG_NAME_2);
+        assertThat(revision, is(headRevision));
+    }
+
+    @Test
+    public void testGetRevisionContainingBranch_OlderName() throws Exception {
+        Revision revision = gitUtils.getRevisionContainingBranch("origin/" + OLDER_BRANCH_NAME);
+        assertThat(revision, is(priorRevision));
+    }
+
+    /* Tags are searched in getRevisionContainingBranch beginning with 3.2.0 */
+    @Test
+    public void testGetRevisionContainingBranch_UseTagNamePrior1() throws Exception {
+        Revision revision = gitUtils.getRevisionContainingBranch("refs/tags/" + PRIOR_TAG_NAME_1);
+        assertThat(revision, is(priorRevision));
+    }
+
+    /* Tags are searched in getRevisionContainingBranch beginning with 3.2.0 */
+    @Test
+    public void testGetRevisionContainingBranch_UseTagNamePrior2() throws Exception {
+        Revision revision = gitUtils.getRevisionContainingBranch("refs/tags/" + PRIOR_TAG_NAME_2);
+        assertThat(revision, is(priorRevision));
+    }
+
+    @Test
+    public void testGetRevisionContainingBranch_InvalidBranchName() throws Exception {
+        Revision revision = gitUtils.getRevisionContainingBranch("origin/not-a-valid-branch-name");
+        assertThat(revision, is(nullValue(Revision.class)));
+    }
+
+    @Test
+    public void testGetRevisionContainingBranch_InvalidTagName() throws Exception {
+        Revision revision = gitUtils.getRevisionContainingBranch("ref/tags/not-a-valid-tag-name");
+        assertThat(revision, is(nullValue(Revision.class)));
+    }
+
+    @Test
+    public void testGetRevisionForSHA1() throws Exception {
+        Revision revision = gitUtils.getRevisionForSHA1(headId);
+        assertThat(revision, is(headRevision));
+    }
+
+    @Test
+    public void testGetRevisionForSHA1PriorRevision() throws Exception {
+        Revision revision = gitUtils.getRevisionForSHA1(priorHeadId);
+        assertThat(revision, is(priorRevision));
     }
 
     @Test
     public void testFilterTipBranches() throws Exception {
         Collection<Revision> multiRevisionList = new ArrayList<>();
-        multiRevisionList.add(remoteHeadRevision);
-        multiRevisionList.add(remotePriorRevision);
+        multiRevisionList.add(priorRevision);
+        multiRevisionList.add(headRevision);
         Collection<Revision> filteredRevisions = new ArrayList<>();
-        filteredRevisions.add(remoteHeadRevision);
+        filteredRevisions.add(headRevision);
         List<Revision> result = gitUtils.filterTipBranches(multiRevisionList);
         assertThat(result, is(filteredRevisions));
-    }
-
-    @Test
-    public void testFilterTipBranchesNoRemovals() throws Exception {
-        Collection<Revision> headRevisionList = new ArrayList<>();
-        headRevisionList.add(remoteHeadRevision);
-        List<Revision> result = gitUtils.filterTipBranches(headRevisionList);
-        assertThat(result, is(headRevisionList));
-    }
-
-    @Test
-    public void testFilterTipBranchesNoRemovalsNonTip() throws Exception {
-        Collection<Revision> priorRevisionList = new ArrayList<>();
-        priorRevisionList.add(remotePriorRevision);
-        List<Revision> result = gitUtils.filterTipBranches(priorRevisionList);
-        assertThat(result, is(priorRevisionList));
-    }
-
-    @Test
-    public void testFixupNames() {
-        String[] names = {"origin", "origin2", null, "", null};
-        String[] urls = {
-            "git://github.com/jenkinsci/git-plugin.git",
-            "git@github.com:jenkinsci/git-plugin.git",
-            "https://github.com/jenkinsci/git-plugin",
-            "https://github.com/jenkinsci/git-plugin.git",
-            "ssh://github.com/jenkinsci/git-plugin.git"
-        };
-        String[] expected = {"origin", "origin2", "origin1", "origin3", "origin4"};
-        String[] actual = GitUtils.fixupNames(names, urls);
-        assertArrayEquals(expected, actual);
     }
 }
